tests/selenium-tests/requirements.txt
allure-behave==2.5.3
allure-python-commons==2.5.3
attrs==18.2.0
behave==1.2.6
delayed_assert==0.3.2
pandas>=1.0
parse==1.8.4
parse-type==0.4.2
pip>=19.0.3
pluggy==0.8.0
PyHamcrest==1.9.0
python-dateutil==2.8.1
selenium==3.14.0
selenium-wire==0.8.0
setuptools==40.2.0
six==1.11.0
urllib3==1.23
wheel==0.31.1
xlrd==1.2.0
requests-futures==0.9.7
